@@ -1,74 +1,59 @@
 {
+  "name": "espiritismo",
+  "version": "1.0.0",
   "main": "node_modules/expo/AppEntry.js",
-  "scripts": {
-    "start": "node_modules/.bin/expo start",
-    "android": "node_modules/.bin/expo start --android",
-    "ios": "node_modules/.bin/expo start --ios",
-    "eject": "node_modules/.bin/expo eject",
-    "test-wait": "node node_modules/jest/bin/jest.js --watchAll",
-    "test": "node node_modules/jest/bin/jest.js --detectOpenHandles"
-  },
-  "jest": {
-    "preset": "jest-expo",
-    "transformIgnorePatterns": [
-      "node_modules/(?!redux-persist)/"
-    ]
-  },
   "importSort": {
     ".js, .jsx, .ts, .tsx": {
       "style": "module"
     }
   },
+  "scripts": {
+    "start": "expo start",
+    "android": "expo start --android",
+    "ios": "expo start --ios",
+    "web": "expo start --web"
+  },
   "dependencies": {
-    "@expo/react-native-action-sheet": "^3.5.0",
-    "@expo/samples": "2.1.1",
-    "@expo/vector-icons": "^10.0.0",
-    "@react-native-community/async-storage": "^1.7.1",
-    "@unpourtous/react-native-search-list": "https://github.com/LVCarnevalli/react-native-search-list/tarball/master",
-    "expo": "^37.0.0",
-    "expo-asset": "~8.1.5",
-    "expo-firebase-analytics": "~2.1.1",
-    "expo-firebase-core": "~1.0.0",
-    "expo-font": "~8.1.0",
-    "firebase": "7.9.0",
+    "@expo/react-native-action-sheet": "^4.0.1",
+    "@expo/vector-icons": "^13.0.0",
+    "@react-native-async-storage/async-storage": "1.17.11",
+    "@react-navigation/native": "^6.1.6",
+    "@react-navigation/stack": "^6.3.16",
+    "expo": "~48.0.15",
+    "expo-asset": "~8.9.1",
+    "expo-font": "~11.1.1",
+    "expo-splash-screen": "^0.18.2",
+    "firebase": "^9.21.0",
     "lodash": "^4.17.11",
-    "logkitty": "0.7.1",
-    "react": "16.9.0",
-    "react-dom": "16.9.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-37.0.1.tar.gz",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-native": "^0.71.7",
     "react-native-circular-progress": "^1.3.6",
-    "react-native-dotenv": "^0.2.0",
-    "react-native-easy-toast": "^1.1.0",
-    "react-native-elements": "^0.19.1",
+    "react-native-dotenv": "^3.4.8",
+    "react-native-easy-toast": "^2.3.0",
+    "react-native-elements": "^3.4.3",
+    "react-native-gesture-handler": "^2.9.0",
     "react-native-global-props": "^1.1.5",
-    "react-native-ignore-warnings": "^0.1.1",
-    "react-native-loading-spinner-overlay": "^1.1.0",
-    "react-native-modal": "^11.5.4",
+    "react-native-loading-spinner-overlay": "^3.0.1",
+    "react-native-modal": "^13.0.1",
     "react-native-parallax-swiper": "^1.1.7",
+    "react-native-safe-area-context": "4.5.0",
+    "react-native-screens": "^3.20.0",
     "react-native-settings-list": "^1.8.0",
-    "react-native-svg": "11.0.1",
+    "react-native-svg": "13.4.0",
     "react-native-swiper": "1.6.0-dev",
-    "react-native-web": "^0.11.7",
-    "react-navigation": "^2.18.2",
+    "react-native-web": "~0.18.1",
+    "react-navigation": "4.4.4",
+    "react-navigation-stack": "^2.10.4",
     "react-process-string": "^1.2.0",
-    "react-redux": "^6.0.0",
-    "redux": "^4.0.1",
+    "react-redux": "^8.0.5",
+    "redux": "^4.2.1",
     "redux-logger": "^3.0.6",
-    "redux-persist": "5.10.0",
-    "redux-thunk": "^2.3.0"
+    "redux-persist": "^6.0.0",
+    "redux-thunk": "^2.4.2"
   },
   "devDependencies": {
-    "babel-eslint": "^10.0.3",
-    "babel-preset-expo": "^8.1.0",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.18.3",
-    "import-sort-style-module": "^6.0.0",
-    "jest-expo": "^37.0.0",
-    "prettier": "^1.19.1",
-    "prettier-plugin-import-sort": "^0.0.3"
+    "@babel/core": "^7.20.0"
   },
   "private": true
 }
