@@ -1,7 +1,7 @@
 import { Feather } from '@expo/vector-icons';
 import React from 'react';
 import { View } from 'react-native';
-import { HeaderBackButton } from 'react-navigation';
+import { HeaderBackButton } from 'react-navigation-stack';
 
 import Layout from '../../constants/Layout';
 import PrayerScreen from '../../screens/Prayer';
@@ -9,6 +9,7 @@ import PrayerScreen from '../../screens/Prayer';
 const PrayerStack = {
   screen: PrayerScreen,
   navigationOptions: ({ navigation }) => ({
+    gestureEnabled: false,
     headerTitle: 'PRECES',
     headerTitleStyle: {
       fontWeight: '400',
@@ -23,12 +24,13 @@ const PrayerStack = {
       elevation: 0,
       borderBottomWidth: 0,
     },
-    headerRight: <View></View>,
-    headerLeft: (
+    headerRight: () => <View></View>,
+    headerLeft: () => (
       <HeaderBackButton
+        labelVisible={false}
         onPress={() => navigation.navigate('Menu')}
         backImage={
-          <Feather
+          () => <Feather
             name="home"
             size={30}
             style={{ color: '#FFFFFF', paddingLeft: Layout.headerPadding }}
