@@ -1,11 +1,13 @@
 import { Feather } from '@expo/vector-icons';
 import React from 'react';
-import { HeaderBackButton } from 'react-navigation';
+import { HeaderBackButton } from 'react-navigation-stack';
 
 import Layout from '../../constants/Layout';
 import QuestionScreen from '../../screens/Question';
 
-const options = ({ navigation }) => ({
+const options = ({ navigation, route }) => ({
+  title: route.params ? route.params.title : '',
+  gestureEnabled: false,
   headerTitleStyle: {
     fontWeight: '400',
     textAlign: 'center',
@@ -13,10 +15,11 @@ const options = ({ navigation }) => ({
     flex: 1,
     fontSize: 18,
   },
-  headerLeft: (
+  headerLeft: () => (
     <HeaderBackButton
+      labelVisible={false}
       onPress={() => navigation.navigate('Menu')}
-      backImage={<Feather name="home" size={30} style={{ paddingLeft: Layout.headerPadding }} />}
+      backImage={() => <Feather name="home" size={30} style={{ paddingLeft: Layout.headerPadding }} />}
     />
   ),
 });
