@@ -1,6 +1,6 @@
 import React from 'react';
 import { View } from 'react-native';
-import { HeaderBackButton } from 'react-navigation';
+import { HeaderBackButton } from 'react-navigation-stack';
 
 import HistoryScreen from '../../../screens/kardec/History';
 
@@ -8,8 +8,8 @@ const HistoryStack = {
   screen: HistoryScreen,
   navigationOptions: ({ navigation }) => ({
     headerTitle: 'Quem foi Kardec?',
-    headerLeft: (
-      <HeaderBackButton onPress={() => navigation.navigate('Kardec')} tintColor={'#000'} />
+    headerLeft: () => (
+      <HeaderBackButton labelVisible={false} onPress={() => navigation.navigate('Kardec')} tintColor={'#000'} />
     ),
     headerTintColor: '#000',
     headerTitleStyle: {
@@ -17,7 +17,7 @@ const HistoryStack = {
       textAlign: 'center',
       flex: 1,
     },
-    headerRight: <View></View>,
+    headerRight: () => <View></View>,
   }),
 };
 
