@@ -1,6 +1,6 @@
 import React from 'react';
 import { View } from 'react-native';
-import { HeaderBackButton } from 'react-navigation';
+import { HeaderBackButton } from 'react-navigation-stack';
 
 import SpiritismScreen from '../../../screens/kardec/Spiritism';
 
@@ -8,8 +8,8 @@ const SpiritismStack = {
   screen: SpiritismScreen,
   navigationOptions: ({ navigation }) => ({
     headerTitle: 'O que é Espiritismo?',
-    headerLeft: (
-      <HeaderBackButton onPress={() => navigation.navigate('Kardec')} tintColor={'#000'} />
+    headerLeft: () => (
+      <HeaderBackButton labelVisible={false} onPress={() => navigation.navigate('Kardec')} tintColor={'#000'} />
     ),
     headerTintColor: '#000',
     headerTitleStyle: {
@@ -17,7 +17,7 @@ const SpiritismStack = {
       textAlign: 'center',
       flex: 1,
     },
-    headerRight: <View></View>,
+    headerRight: () => <View></View>,
   }),
 };
 
